Worker application form updated
I have changed my javascript in this updated "worker application form". When a user enters nothing in the first text box. They will be prompted by an alert box to enter a real value. The length property returns the length of a value that is entered. If nothing is entered, the alert box will keep on prompting for a value until a value is entered.

The else keyword used here implies that if "otherwise", that is if a value has been entered then the block of code ends here and the string in the alert box is shown on the screen.